+++ packages/vorbis_rs/src/common/vorbis.rs
@@ -12,7 +12,7 @@ use std::num::{NonZeroU32, NonZeroU8};
 use std::os::raw::{c_int, c_long};
 use std::ptr;
 
-use crate::common::{assume_init_box, vorbis_error_code_to_string, VorbisError};
+use crate::common::{assume_init_box, VorbisError};
 
 /// A high-level abstraction for a Vorbis stream information struct.
 pub struct VorbisInfo {