+++ packages/vorbis_rs/src/encoder/encoder_impl.rs
@@ -1,6 +1,4 @@
-use crate::common::{
-	vorbis_error_code_to_string, OggPacket, OggStream, VorbisComments, VorbisError, VorbisInfo
-};
+use crate::common::{OggPacket, OggStream, VorbisComments, VorbisError, VorbisInfo};
 use crate::encoder::{VorbisBitrateManagementStrategy, VorbisEncodingState};
 
 use std::borrow::Cow;